requirements.txt
sqlalchemy==2.0.25
psycopg2==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
SQLAlchemy-Utils==0.41.1
fastapi==0.109.0
uvicorn==0.26.0
pydantic==2.5.3
pytest==7.4.4
requests==2.31.0
pytest-asyncio==0.23.3